fix(Development): make development base neutral on decimal rounded value (#649)
If value is 0.01 but decimal is 1 rendered value becomes 0.0. However,
since the development is still positive the Development component
renders 0.0 in the positive styling. This commit changes this behaviour
and makes the development look at the rendered value rather than the
original value, aligning the design and the displayed value.